Fix types missmatch warnings

+++ src/crypto/Lyra2Z/Lyra2.c
@@ -60,10 +60,10 @@ int LYRA2(void *K, uint64_t kLen, const void *pwd, uint64_t pwdlen, const void *
     //Tries to allocate enough space for the whole memory matrix
 
 
-    const int64_t ROW_LEN_INT64 = BLOCK_LEN_INT64 * nCols;
-    const int64_t ROW_LEN_BYTES = ROW_LEN_INT64 * 8;
+    const uint64_t ROW_LEN_INT64 = BLOCK_LEN_INT64 * nCols;
+    const uint64_t ROW_LEN_BYTES = ROW_LEN_INT64 * 8;
 
-    i = (int64_t) ((int64_t) nRows * (int64_t) ROW_LEN_BYTES);
+    i = nRows * ROW_LEN_BYTES;
     uint64_t *wholeMatrix = malloc(i);
     if (wholeMatrix == NULL) {
       return -1;
@@ -160,15 +160,14 @@ int LYRA2(void *K, uint64_t kLen, const void *pwd, uint64_t pwdlen, const void *
       window *= 2; //doubles the size of the re-visitation window
       gap = -gap; //inverts the modifier to the step
     }
-
-    } while (row < nRows);
+    } while (((uint64_t)row) < nRows);
     //==========================================================================/
 
     //============================ Wandering Phase =============================//
     row = 0; //Resets the visitation to the first row of the memory matrix
     for (tau = 1; tau <= timeCost; tau++) {
         //Step is approximately half the number of all rows of the memory matrix for an odd tau; otherwise, it is -1
-        step = (tau % 2 == 0) ? -1 : nRows / 2 - 1;
+        step = (tau % 2 == 0) ? (int64_t)-1 : (int64_t)nRows / 2 - 1;
         do {
         //Selects a pseudorandom index row*
         //------------------------------------------------------------------------------------------
@@ -229,10 +228,10 @@ int LYRA2_old(void *K, uint64_t kLen, const void *pwd, uint64_t pwdlen, const vo
     //Tries to allocate enough space for the whole memory matrix
 
 
-    const int64_t ROW_LEN_INT64 = BLOCK_LEN_INT64 * nCols;
-    const int64_t ROW_LEN_BYTES = ROW_LEN_INT64 * 8;
+    const uint64_t ROW_LEN_INT64 = BLOCK_LEN_INT64 * nCols;
+    const uint64_t ROW_LEN_BYTES = ROW_LEN_INT64 * 8;
 
-    i = (int64_t) ((int64_t) nRows * (int64_t) ROW_LEN_BYTES);
+    i = nRows * ROW_LEN_BYTES;
     uint64_t *wholeMatrix = malloc(i);
     if (wholeMatrix == NULL) {
       return -1;
@@ -329,15 +328,14 @@ int LYRA2_old(void *K, uint64_t kLen, const void *pwd, uint64_t pwdlen, const vo
       window *= 2; //doubles the size of the re-visitation window
       gap = -gap; //inverts the modifier to the step
     }
-
-    } while (row < nRows);
+    } while (((uint64_t)row) < nRows);
     //==========================================================================/
 
     //============================ Wandering Phase =============================//
     row = 0; //Resets the visitation to the first row of the memory matrix
     for (tau = 1; tau <= timeCost; tau++) {
         //Step is approximately half the number of all rows of the memory matrix for an odd tau; otherwise, it is -1
-        step = (tau % 2 == 0) ? -1 : nRows / 2 - 1;
+        step = (tau % 2 == 0) ? (int64_t)-1 : (int64_t)nRows / 2 - 1;
         do {
         //Selects a pseudorandom index row*
         //------------------------------------------------------------------------------------------
